Rerun tests with new setup while making tables comparable
I wanted to check how comparable the new backend is compared to what already exists. There are a few settings that might change, but this is to show that the results are very comparable between the two. There is some minor spacing changes when dealing with multicolumn objects, it is otherwise capable of producing very similar tables.