Fix Tests
After Julia 1.9, there is a warning if a test overwrites a function, which is how the defaults work. For the user, this is not the case, just within the tests. It also appears that changing those defaults also impacts other tests, so the tests for them are now more isolated and reverted right after.